+++ packages/i18nify-js/src/modules/geo/getFlagOfCountry.ts
@@ -1,7 +1,7 @@
-import { CountryCodeType } from '../types';
+import { CountryCodeType, GetFlagReturnType } from '../types';
 import { withErrorBoundary } from '../../common/errorBoundary';
 import { LIST_OF_ALL_COUNTRIES } from './data/listOfAllCountries';
-import { FLAG_BASE_PATH } from './constants';
+import { FLAG_4X3_BASE_PATH, FLAG_BASE_PATH } from './constants';
 
 /**
  * Retrieves the URL for the flag of a given country.
@@ -15,11 +15,15 @@ import { FLAG_BASE_PATH } from './constants';
  * @returns The URL of the flag image for the given country code.
  * @throws {Error} If the country code is not in the list of valid country codes.
  */
-const getFlagOfCountry = (countryCode: CountryCodeType): string => {
+const getFlagOfCountry = (_countryCode: CountryCodeType): GetFlagReturnType => {
+  const countryCode = _countryCode.toUpperCase() as CountryCodeType;
   if (!LIST_OF_ALL_COUNTRIES.includes(countryCode)) {
     throw new Error(`Invalid country code: ${countryCode}`);
   }
-  return `${FLAG_BASE_PATH}/${countryCode.toLowerCase()}.svg`;
+  return {
+    original: `${FLAG_BASE_PATH}/${countryCode}.svg`,
+    '4X3': `${FLAG_4X3_BASE_PATH}/${countryCode}.svg`,
+  };
 };
 
 export default withErrorBoundary<typeof getFlagOfCountry>(getFlagOfCountry);