+++ packages/i18nify-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @razorpay/i18nify-js
 
+## 1.8.0
+
+### Minor Changes
+
+- 33a66ca: Flag service updated to return unpkg url
+
+## 1.7.0
+
+### Minor Changes
+
+- 05e4574: Flags deployed on unpkg
+
 ## 1.6.1
 
 ### Patch Changes